Exercicios_Operadores

+++ Exercicio_4/Celsius_Fahrenheit/Program.cs
@@ -1,8 +1,8 @@
-﻿int celsius = 0;
-int fahrenheit = 0;
+﻿double celsius = 0;
+double fahrenheit = 0;
 
 Console.Write("Digite o valor em °C: ");
-celsius = Convert.ToInt32(Console.ReadLine());
+celsius = Convert.ToDouble(Console.ReadLine());
 
 fahrenheit = (9 * celsius + 160) / 5;
